feat: nix build packages
The current strategy with bun is to have a FOD of node_modules and then
build the actual app in a separate derivation. We use the bun
http-server module to serve the static files, and I'm sure we can come
up with a more secure strategy later with options.

This involved using adapter-static because we have a static site.
Thankfully everything still works!